+++ src/dpsprep/concurrency/message.py
@@ -1,4 +1,3 @@
-import enum
 import logging
 from dataclasses import dataclass
 
@@ -13,15 +12,25 @@ class LogRecordWorkerMessage:
     record: logging.LogRecord
 
 
-class TaskDoneType(enum.IntEnum):
-    TEXT = enum.auto()
-    IMAGE = enum.auto()
-
-
 @dataclass(frozen=True)
-class TaskDoneWorkerMessage:
-    type: TaskDoneType
+class ImagePageProcessedMessage:
     page: int
 
 
-WorkerMessage = ExceptionWorkerMessage | LogRecordWorkerMessage | TaskDoneWorkerMessage
+@dataclass(frozen=True)
+class TextPageProcessedMessage:
+    page: int
+
+
+@dataclass(frozen=True)
+class TextLayerAlreadyProcessedMessage:
+    pass
+
+
+WorkerMessage = (
+    ExceptionWorkerMessage |
+    LogRecordWorkerMessage |
+    ImagePageProcessedMessage |
+    TextPageProcessedMessage |
+    TextLayerAlreadyProcessedMessage
+)