+++ src/dpsprep/outline/__init__.py
@@ -1,2 +1,2 @@
 from .outline import extract_outline_as_pdfdict
-from .text import extract_text_as_fpdf
+from .text import TextExtractionTracker, extract_text_as_fpdf